xml/parsing/manakai/tree-1.dat

#data
<a/>
#errors
1;1;no XML decl;;;s
#document
| <{} a>

#data
<a></a>
#errors
1;1;no XML decl;;;s
#document
| <{} a>

#data
<a><b></a>
#errors
1;1;no XML decl;;;s
1;7;not closed before ancestor end tag;b;a;m
#document
| <{} a>
|   <{} b>

#data
<a><b></a>a
#errors
1;1;no XML decl;;;s
1;7;not closed before ancestor end tag;b;a;m
1;11;text outside of root element;;;m
#document
| <{} a>
|   <{} b>

#data
<a/><b/>
#errors
1;1;no XML decl;;;s
1;5;second root element;;;m
1;5;nestc;b;;m
#document
| <{} a>
